+++ src/Braintree/PaymentMethodGateway.cs
@@ -64,7 +64,7 @@ namespace Braintree
                 throw new NotFoundException();
 
             var response = new NodeWrapper(service.Get(service.MerchantPath() + "/payment_methods/any/" + token));
-            return ExtractPaymentMethodFromResponse(response);
+            return PaymentMethodParser.ParsePaymentMethod(response, gateway);
         }
 
         public async Task<PaymentMethod> FindAsync(string token)
@@ -73,7 +73,7 @@ namespace Braintree
                 throw new NotFoundException();
 
             var response = new NodeWrapper(await service.GetAsync(service.MerchantPath() + "/payment_methods/any/" + token).ConfigureAwait(false));
-            return ExtractPaymentMethodFromResponse(response);
+            return PaymentMethodParser.ParsePaymentMethod(response, gateway);
         }
 
         public Result<PaymentMethodNonce> Grant(string token, PaymentMethodGrantRequest request)
@@ -142,54 +142,6 @@ namespace Braintree
             else
             {
                 return new ResultImpl<UnknownPaymentMethod>(response, gateway);
-            }
-        }
-
-        private PaymentMethod ExtractPaymentMethodFromResponse(NodeWrapper response)
-        {
-            if (response.GetName() == "paypal-account")
-            {
-                return new PayPalAccount(response, gateway);
-            }
-            else if (response.GetName() == "us-bank-account")
-            {
-                return new UsBankAccount(response);
-            }
-            else if (response.GetName() == "credit-card")
-            {
-                return new CreditCard(response, gateway);
-            }
-            else if (response.GetName() == "apple-pay-card")
-            {
-                return new ApplePayCard(response, gateway);
-            }
-            else if (response.GetName() == "android-pay-card")
-            {
-                return new AndroidPayCard(response, gateway);
-            }
-            else if (response.GetName() == "coinbase-account")
-            {
-                return new CoinbaseAccount(response, gateway);
-            }
-            else if (response.GetName() == "venmo-account")
-            {
-                return new VenmoAccount(response, gateway);
-            }
-            else if (response.GetName() == "visa-checkout-card")
-            {
-                return new VisaCheckoutCard(response, gateway);
-            }
-            else if (response.GetName() == "masterpass-card")
-            {
-                return new MasterpassCard(response, gateway);
-            }
-            else if (response.GetName() == "samsung-pay-card")
-            {
-                return new SamsungPayCard(response, gateway);
-            }
-            else
-            {
-                return new UnknownPaymentMethod(response);
             }
         }
     }