+++ test/Braintree.Tests.Integration/DocumentUploadIntegrationTest.cs
@@ -25,6 +25,12 @@ namespace Braintree.Tests.Integration
 #endif
 
 #if netcore
+        private static string TOO_LONG_FILE_PATH = "../../../../../test/fixtures/too_long.pdf";
+#else
+        private static string TOO_LONG_FILE_PATH = "test/fixtures/too_long.pdf";
+#endif
+
+#if netcore
         private static string MALFORMED_FILE_PATH = "../../../../../test/fixtures/malformed_pdf.pdf";
 #else
         private static string MALFORMED_FILE_PATH = "test/fixtures/malformed_pdf.pdf";
@@ -292,6 +298,44 @@ namespace Braintree.Tests.Integration
             Assert.IsFalse(result.IsSuccess());
 
             Assert.AreEqual(ValidationErrorCode.DOCUMENT_UPLOAD_FILE_IS_TOO_LARGE, result.Errors.ForObject("DocumentUpload").OnField("File")[0].Code);
+        }
+#if net452
+            ).GetAwaiter().GetResult();
+        }
+#endif
+
+        [Test]
+        public void Create_returnsErrorWhenFileIsOver50Pages()
+        {
+            FileStream fs = new FileStream(TOO_LONG_FILE_PATH, FileMode.Open, FileAccess.Read);
+            DocumentUploadRequest request = new DocumentUploadRequest();
+            request.File = fs;
+            request.DocumentKind = DocumentUploadKind.EVIDENCE_DOCUMENT;
+            var result = gateway.DocumentUpload.Create(request);
+
+            Assert.IsFalse(result.IsSuccess());
+
+            Assert.AreEqual(ValidationErrorCode.DOCUMENT_UPLOAD_FILE_IS_TOO_LONG, result.Errors.ForObject("DocumentUpload").OnField("File")[0].Code);
+        }
+
+        [Test]
+#if netcore
+        public async Task CreateAsync_returnsErrorWhenFileIsOver50Pages()
+#else
+        public void CreateAsync_returnsErrorWhenFileIsOver50Pages()
+        {
+            Task.Run(async () =>
+#endif
+        {
+            FileStream fs = new FileStream(TOO_LONG_FILE_PATH, FileMode.Open, FileAccess.Read);
+            DocumentUploadRequest request = new DocumentUploadRequest();
+            request.File = fs;
+            request.DocumentKind = DocumentUploadKind.EVIDENCE_DOCUMENT;
+            var result = await gateway.DocumentUpload.CreateAsync(request);
+
+            Assert.IsFalse(result.IsSuccess());
+
+            Assert.AreEqual(ValidationErrorCode.DOCUMENT_UPLOAD_FILE_IS_TOO_LONG, result.Errors.ForObject("DocumentUpload").OnField("File")[0].Code);
         }
 #if net452
             ).GetAwaiter().GetResult();