+++ test/Braintree.Tests.Integration/PaymentMethodNonceIntegrationTest.cs
@@ -84,9 +84,24 @@ namespace Braintree.Tests.Integration
             Assert.AreEqual(foundNonce.Nonce, nonce);
             Assert.AreEqual(foundNonce.Type, "CreditCard");
             Assert.IsNotNull(foundNonce.Details);
+            Assert.AreEqual(foundNonce.Details.Bin, "401288");
             Assert.AreEqual(foundNonce.Details.CardType, "Visa");
             Assert.AreEqual(foundNonce.Details.LastTwo, "81");
             Assert.AreEqual(foundNonce.Details.LastFour, "1881");
+        }
+
+
+        [Test]
+        public void Find_ExposesDetailsForVenmoNonce()
+        {
+            string nonce = "fake-venmo-account-nonce";
+            PaymentMethodNonce foundNonce = gateway.PaymentMethodNonce.Find(nonce);
+            Assert.IsNotNull(foundNonce);
+            Assert.AreEqual(foundNonce.Nonce, nonce);
+            Assert.IsNotNull(foundNonce.Details);
+            Assert.AreEqual(foundNonce.Details.Username, "venmojoe");
+            Assert.AreEqual(foundNonce.Details.LastTwo, "99");
+            Assert.AreEqual(foundNonce.Details.VenmoUserId, "Venmo-Joe-1");
         }
 
         [Test]